% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_builder.r
\docType{methods}
\name{buildRiskModelForDateRange}
\alias{buildRiskModelForDateRange}
\alias{buildRiskModelForDateRange,VirtualRiskModelBuilder,Date,Date-method}
\title{Build risk model data for specific date range}
\usage{
buildRiskModelForDateRange(object, date_start, date_end)


  \S4method{buildRiskModelForDateRange}{VirtualRiskModelBuilder,Date,Date}(object,
  date_start, date_end)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelBuilder'.}

\item{date_start}{"Date" start date of computation}

\item{date_end}{"Date" end date of the model}
}
\value{
\code{status} 'logical', TRUE if model was updated, FALSE otherwise
}
\description{
Builds risk model data for specific risk model and specific date range
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualRiskModelBuilder,date_start = Date,date_end = Date}: Build risk model data for specific date range

Builds risk model data for specific risk model and specific date range
}}

