% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{composite_model}
\alias{composite_model}
\title{Compute composite model for blocks of factors}
\usage{
composite_model(stock_returns_on_date, factor_blocks)
}
\arguments{
\item{stock_returns_on_date}{"data.frame" with stock returns for date}

\item{factor_blocks}{list containing "data.frame" with returns of factors within a block}
}
\value{
\code{x_sectional_model} list with two components :\cr
\code{x_sectional_model$log_implied_rtn} - "data.frame" with log of implied factor returns\cr
\code{x_sectional_model$log_implied_rtn} - "data.frame" with log of stock residual returns
}
\description{
Point in time composite model computed against preceeding block residuals:
Single cross sectional regression model is built for each factor block using the
preceeding block residuals
Expects stock return frame with columns Instrument and Return
Expects stock weights in frame with columns for Instrument and Weight
Factor blocks is a list of dataframes with one column for Instrument and a column for each factor
}

