% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{factor_correlation}
\alias{factor_correlation}
\title{Compute the factor correlation matrix from the implied factor returns}
\usage{
factor_correlation(implied_fct_rtns)
}
\arguments{
\item{implied_fct_rtns}{data.frame with implied factor returns}
}
\value{
\code{factor_cor} "data.frame" with factor correlation
}
\description{
Compute the factor correlation matrix from the implied factor returns
}

