% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelBetaEstimatorType,VirtualBetaEstimator-method}
\alias{getRiskModelBetaEstimator}
\alias{getRiskModelBetaEstimator,VirtualRiskModel-method}
\alias{getRiskModelBetaEstimatorType,VirtualBetaEstimator-method}
\title{Get Risk Model Beta Estimator}
\usage{
\S4method{getRiskModelBetaEstimatorType}{VirtualBetaEstimator}(object)

getRiskModelBetaEstimator(object)

\S4method{getRiskModelBetaEstimator}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{beta_model} 'VirtualBetaEstimator', object capturing beta model properties

\code{beta_estimator} 'VirtualBetaEstimator', object capturing beta model properties
}
\description{
Returns name of Risk Model Beta Estimator object
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualBetaEstimator}: Get Risk Model Name

Returns name of Risk Model

\item \code{VirtualRiskModel}: Get Risk Model Beta Estimator Object

Returns Risk Model Beta Estimator Object
}}

