% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelBetaEstimatorType}
\alias{getRiskModelBetaEstimatorType}
\title{Get Estimator method Name}
\usage{
getRiskModelBetaEstimatorType(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\description{
Returns name of Betas estimator
}

