% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelCommodityFactorNames}
\alias{getRiskModelCommodityFactorNames}
\alias{getRiskModelCommodityFactorNames,VirtualRiskModel-method}
\title{Get Risk Model Commodity Factor Names}
\usage{
getRiskModelCommodityFactorNames(object)

\S4method{getRiskModelCommodityFactorNames}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_factors} 'character', vector of factor names
}
\description{
Returns names of factors included in the model
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Commodity Factor Names

Returns names of sector factors included in the model
}}

