% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelCurrencyFactorNames}
\alias{getRiskModelCurrencyFactorNames}
\alias{getRiskModelCurrencyFactorNames,VirtualRiskModel-method}
\title{Get Risk Model Currency Factor Names}
\usage{
getRiskModelCurrencyFactorNames(object)

\S4method{getRiskModelCurrencyFactorNames}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_factors} 'character', vector of factor names
}
\description{
Returns names of factors included in the model
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Currency Factor Names

Returns names of sector factors included in the model
}}

