% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelFactorGroups}
\alias{getRiskModelFactorGroups}
\alias{getRiskModelFactorGroups,VirtualRiskModel-method}
\title{Set Risk Model User defined Factor Groups}
\usage{
getRiskModelFactorGroups(object)

\S4method{getRiskModelFactorGroups}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{factor_groups} list with character objects specifying factor grouppings
}
\description{
Set Risk Model User defined Factor Groups
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: get Risk Model User defined Factor Groups
}}

