% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelFactorNames,VirtualRiskModelHandler-method}
\alias{getRiskModelFactorNames,VirtualRiskModelHandler-method}
\title{Get Risk Model Factor Names}
\usage{
\S4method{getRiskModelFactorNames}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{model_factors} 'character', vector of factor names
}
\description{
Returns names of factors included in the model
}

