% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelBetaEstimatorType,VirtualBetaEstimator-method}
\alias{getRiskModelBetaEstimatorType,VirtualBetaEstimator-method}
\alias{getRiskModellBetaEstimator}
\alias{getRiskModellBetaEstimator,VirtualRiskModel-method}
\title{Get Risk Model Name}
\usage{
\S4method{getRiskModelBetaEstimatorType}{VirtualBetaEstimator}(object)

getRiskModellBetaEstimator(object)

\S4method{getRiskModellBetaEstimator}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{beta_model} 'VirtualBetaEstimator', object capturing beta model properties

\code{beta_model} 'VirtualBetaEstimator', object capturing beta model properties
}
\description{
Returns name of Risk Model
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualBetaEstimator}: Get Risk Model Name

Returns name of Risk Model

\item \code{VirtualRiskModel}: Get Risk Model Name

Returns name of Risk Model
}}

