% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelLookback}
\alias{getRiskModelLookback}
\alias{getRiskModelLookback,VirtualRiskModel-method}
\title{Get Risk Model Lookback}
\usage{
getRiskModelLookback(object)

\S4method{getRiskModelLookback}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{lookback} 'integer'
}
\description{
Returns model lookback value in days. Lookback is number of
calendar days over which model is calculated.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Lookback

Returns model lookback value in days. Lookback is number of
calendar days over which model is calculated.
}}

