% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelName}
\alias{getRiskModelName}
\alias{getRiskModelName,VirtualRiskModel-method}
\title{Get Risk Model Name}
\usage{
getRiskModelName(object)

\S4method{getRiskModelName}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_name} 'character', name of the model.
}
\description{
Returns name of Risk Model
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Name

Returns name of Risk Model
}}

