% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelPrefix}
\alias{getRiskModelPrefix}
\alias{getRiskModelPrefix,VirtualRiskModel-method}
\title{Get Risk Model Prefix}
\usage{
getRiskModelPrefix(object)

\S4method{getRiskModelPrefix}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_prefix} 'character', model prefix.
}
\description{
Returns prefix of Risk Model used to construct model store querries
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Prefix

Returns prefix of Risk Model used to construct model store querries
}}

