% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelSectorFactorNames}
\alias{getRiskModelSectorFactorNames}
\alias{getRiskModelSectorFactorNames,VirtualRiskModel-method}
\title{Get Risk Model Sector Factor Names}
\usage{
getRiskModelSectorFactorNames(object)

\S4method{getRiskModelSectorFactorNames}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_factors} 'character', vector of factor names
}
\description{
Returns names of factors included in the model
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Factor Names

Returns names of sector factors included in the model
}}

