% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelSectorFactorReturns}
\alias{getRiskModelSectorFactorReturns}
\alias{getRiskModelSectorFactorReturns,VirtualRiskModel,Date,Date-method}
\title{Get Risk Model Sector Factor returns}
\usage{
getRiskModelSectorFactorReturns(object, date_start, date_end)

\S4method{getRiskModelSectorFactorReturns}{VirtualRiskModel,Date,Date}(object,
  date_start, date_end)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}

\item{date_start}{"Date" start date of requested factor returns}

\item{date_end}{"Date" end date of requested factor returns}
}
\value{
\code{fct_returns} 'data.frame' with "Date" "FactorName" and "Return" columns
}
\description{
Returns data.frame with daily returns of factors included in the model
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualRiskModel,date_start = Date,date_end = Date}: Get Risk Model Sector Factor returns

Returns data.frame with daily returns of factors included in the model
}}

