% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelUserFactorGroups}
\alias{getRiskModelUserFactorGroups}
\alias{getRiskModelUserFactorGroups,VirtualRiskModel-method}
\title{Set Risk Model User defined Factor Groups}
\usage{
getRiskModelUserFactorGroups(object)

\S4method{getRiskModelUserFactorGroups}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{factor_groups} list with character objects specifying user factor grouppings
}
\description{
Set Risk Model User defined Factor Groups
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Set Risk Model User defined Factor Groups
}}

