% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{initialize,VirtualRiskModel-method}
\alias{initialize,VirtualRiskModel-method}
\title{initialize method for VirtualRiskModel}
\usage{
\S4method{initialize}{VirtualRiskModel}(.Object)
}
\description{
initialize method for VirtualRiskModel
}

