% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{instrument_regression}
\alias{instrument_regression}
\title{Perform regressions for one instrument}
\usage{
instrument_regression(ins, all_data)
}
\arguments{
\item{ins}{"integer" instrument id}

\item{all_data}{"data.frame" with all instruments returns}
}
\value{
\code{rval} "data.frame" with all instruments betas in columns named after factor names
}
\description{
function to apply regression for all factors
}

