% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{omit_value}
\alias{omit_value}
\title{Remove values that disrupt regression}
\usage{
omit_value(values_arr)
}
\arguments{
\item{values_arr}{array of values to be cleared}
}
\value{
\code{values_arr} cleared array
}
\description{
Clears array from NA's, INF's and NAN's
}

