% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{pfo_variance_decomposition}
\alias{pfo_variance_decomposition}
\title{Compute portfolio variance decomposition}
\usage{
pfo_variance_decomposition(weight, betas, factor_covariance, columns = NULL)
}
\arguments{
\item{weight}{"data.frame" with portfolio constituents weights}

\item{betas}{"data.frame" with portfolio instrument's betas}

\item{factor_covariance}{"matrix" with factor covariance}

\item{columns}{"character" list of factors to account when computing decomposition.
Default is NULL. If missing or NULL all factors will be used}
}
\description{
This computes the variance of the log returns. This is not identical to the variance of the
returns. However we take the variance of the log returns as equal to the variance of the
returns beacuse the Taylor expansion of the variance of the log returns gives var[log(R)] ~ var[R]/mean[R]^2
and since the mean return will generally be close to one, this is a reasonable approximation.
If the mean return is above 1, then the risk measure will slightly underestimate risk, whereas if
mean return is lower than 1, risks will be slightly over estimated.
}

