% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{.getRiskModelFactorGroups}
\alias{.getRiskModelFactorGroups}
\title{Get Risk Model Factor Groups}
\usage{
.getRiskModelFactorGroups(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\description{
Returns list of factor groups
}

