% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\name{.setRiskModelObject}
\alias{.setRiskModelObject}
\title{Set Risk Model Object}
\usage{
.setRiskModelObject(object, risk_model)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}

\item{risk_model}{object of class "VirtualRiskModel".}
}
\value{
\code{object} object of class 'VirtualRiskModelHandler'.
}
\description{
Private method to set risk_model object slot value
}

