% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{rgr_kernel}
\alias{rgr_kernel}
\title{kernel function to run regression}
\usage{
rgr_kernel(xi, data)
}
\arguments{
\item{xi}{array with x values}

\item{data}{array with values to be regressed}
}
\value{
\code{rval} list with components returned by lm.fit
}
\description{
function to apply regression, but also to remove values that
will cause regression to fail.
using lm.fit because it is faster
}

