% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{data}
\name{risk_model_commodity_factors}
\alias{risk_model_commodity_factors}
\title{List of risk model commodity factors}
\format{An object of class \code{character} of length 1.}
\usage{
risk_model_commodity_factors
}
\description{
List of risk model commodity factors
}
\keyword{datasets}

