% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{data}
\name{risk_model_currency_factors}
\alias{risk_model_currency_factors}
\title{List of risk model currency factors}
\format{An object of class \code{character} of length 16.}
\usage{
risk_model_currency_factors
}
\description{
List of risk model currency factors
}
\keyword{datasets}

