% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{data}
\name{risk_model_market_factors}
\alias{risk_model_market_factors}
\title{List of risk model market factors}
\format{An object of class \code{character} of length 10.}
\usage{
risk_model_market_factors
}
\description{
List of risk model market factors
}
\keyword{datasets}

