% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{data}
\name{risk_model_sector_factors}
\alias{risk_model_sector_factors}
\title{List of risk model sector factors}
\format{An object of class \code{character} of length 18.}
\usage{
risk_model_sector_factors
}
\description{
List of risk model sector factors
}
\keyword{datasets}

