% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{class}
\name{RiskModel.DevelopedEurope.1.2-class}
\alias{RiskModel.DevelopedEurope.1.2-class}
\title{Virtual S4 class for Developed Europe Risk Models.}
\description{
Class implementing parameters for family of Developed
Europe Risk Models version 1.1 excluding HKD and DKK
And robust linear model betas estimator
}
\details{
Inherits from "VirtualRiskModel"
}

