% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{class}
\name{RiskModel.DevelopedEuropePrototype150.1.2-class}
\alias{RiskModel.DevelopedEuropePrototype150.1.2-class}
\title{Concrete S4 class for Developed Europe Risk Model.}
\description{
Class implementing parameters for Prototype Developed
Europe Risk Model with 150 days lookback
and Robust Linear Model Beta estimator
}
\details{
Inherits from "RiskModel.DevelopedEurope.1.1"
}

