% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_builder.r
\docType{class}
\name{RiskModelBuilder-class}
\alias{RiskModelBuilder-class}
\title{Concrete S4 class implementing methods to handle building of
Risk Models historical data}
\description{
Handles building up historical model data and updating
}
\details{
Inherits from VirtualRiskModelBuilder
}

