% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{setRiskModelUserFactorGroups}
\alias{setRiskModelUserFactorGroups}
\alias{setRiskModelUserFactorGroups,VirtualRiskModel,list-method}
\title{Set Risk Model User defined Factor Groups}
\usage{
setRiskModelUserFactorGroups(object, factor_groups)

\S4method{setRiskModelUserFactorGroups}{VirtualRiskModel,list}(object,
  factor_groups)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}

\item{factor_groups}{list with character objects specifying factor grouppings}
}
\value{
\code{object} object of class 'VirtualRiskModel'
}
\description{
Set Risk Model User defined Factor Groups
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualRiskModel,factor_groups = list}: Set Risk Model User defined Factor Groups
}}

