% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{stock_correlation}
\alias{stock_correlation}
\title{Compute stock covariance matrix}
\usage{
stock_correlation(betas, implied_fct_rtns, factor_correlation_matrix,
  risk_model_date)
}
\arguments{
\item{betas}{"data.frame" with factor betas}

\item{implied_fct_rtns}{"data.frame" with implied factor returns}

\item{factor_correlation_matrix}{"data.frame" with factor correlation matrix}

\item{risk_model_date}{"Date" with risk model date}
}
\value{
\code{stock_corr} "matrix" with stock correlation
}
\description{
Use the stock betas and the factor correlation matrix to generate the
stock correlation matrix
}

