% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_builder.r
\docType{methods}
\name{updateRiskModelOnDate}
\alias{updateRiskModelOnDate}
\alias{updateRiskModelOnDate,VirtualRiskModelBuilder,Date,logical,logical-method}
\title{Update risk model on specific date}
\usage{
updateRiskModelOnDate(object, date, force = FALSE, copy_history = FALSE)


  \S4method{updateRiskModelOnDate}{VirtualRiskModelBuilder,Date,logical,logical}(object,
  date = today() - 1, force = FALSE, copy_history = FALSE)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelBuilder'.}

\item{date}{"Date" date of the model for which the entries should be computed}

\item{force}{"logical" should the values be replaced if already existing}

\item{copy_history}{"logical" should we look for some components history if computed earlier}
}
\value{
\code{status} 'logical', TRUE if model was updated, FALSE otherwise
}
\description{
updates risk model up to specified date for month of the date.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualRiskModelBuilder,date = Date,force = logical,copy_history = logical}: Update risk model on specific date

updates risk model up to specified date for month of the date.
}}

