% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{class}
\name{VirtualBetaEstimator-class}
\alias{VirtualBetaEstimator-class}
\title{Virtual S4 class for handling Beta Estimator Properties.}
\description{
Virtual S4 class for handling Beta Estimator Properties.
}
\section{Slots}{

\describe{
\item{\code{est_method_name}}{"character"}
}}

