% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_builder.r
\docType{class}
\name{VirtualRiskModelBuilder-class}
\alias{VirtualRiskModelBuilder-class}
\title{Virtual S4 Class implementing methods to handle building of
Risk Models historical data}
\description{
Handles building up historical model data and updating
}
\details{
Inherits from VirtualRiskModelHandler
}

