% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{class}
\name{VirtualRiskModelHandler-class}
\alias{VirtualRiskModelHandler-class}
\title{Virtual S4 class for handling Risk Model class.}
\description{
This is handler class that is to be inherited
by other classes handling Risk Model Objects
}
\details{
Inherits from "VirtualReferenceObject"
}
\section{Slots}{

\describe{
\item{\code{risk_model}}{"VirtualRiskModel", default is "RiskModel.DevelopedEuropePrototype150"}
}}

