% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_functions.r
\name{x_sectional_model}
\alias{x_sectional_model}
\title{Cross sectional factor return computation}
\usage{
x_sectional_model(stock_return, stock_betas)
}
\arguments{
\item{stock_return}{"data.frame" with stock returns}

\item{stock_betas}{"data.frame" with stock betas}
}
\value{
\code{x_sectional_model} list with two components :\cr
\code{x_sectional_model$log_implied_rtn} - "data.frame" with log of implied factor returns\cr
\code{x_sectional_model$log_implied_rtn} - "data.frame" with log of stock residual returns
}
\description{
Point in time estimation of risk model:
Cross sectional regression of stock returns against raw factor returns
to produce the implied factor returns
Expects stock return frame with columns Instrument and Return
Expects stock betas in a frame with one column for Instrument and a column for each factor
Expects stock weights in frame with columns for Instrument and Weight
}

