% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_update_functions.r
\name{update_risk_model_on_date}
\alias{update_risk_model_on_date}
\title{update missing risk model entries up to date}
\usage{
update_risk_model_on_date(model_prefix, date, lookback = 150, force = FALSE,
  copy_history = FALSE)
}
\arguments{
\item{model_prefix}{"character" name of the model eg. "developed_europe_prototype"}

\item{date}{"Date" date of the model for which the entries should be computed}

\item{lookback}{"integer" number of days over which the model is computed.
These are callendar days therefore the model can actually contain less trading days.}

\item{force}{"logical" should the values be replaced if already existing}

\item{copy_history}{"logical" should we look for some components history in earlier
risk model objectstores}
}
\description{
checks existing risk model data and computes model values
for entries up to specified date
}

