% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{class}
\name{VirtualRiskModel-class}
\alias{VirtualRiskModel-class}
\title{Virtual S4 class for handling Risk Model Properties.}
\description{
Virtual S4 class for handling Risk Model Properties.
}
\section{Slots}{

\describe{
\item{\code{model_prefix}}{"character"}

\item{\code{lookback}}{"integer"}

\item{\code{model_universe}}{"character"}

\item{\code{model_factors}}{"character"}
}}

